+++ src/main/java/frc/robot/commands/Launcher/SemiAutoVersion4.java
@@ -7,29 +7,36 @@
 
 package frc.robot.commands.Launcher;
 
+import edu.wpi.first.networktables.NetworkTable;
+import edu.wpi.first.networktables.NetworkTableEntry;
+import edu.wpi.first.networktables.NetworkTableInstance;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.IndexSpeed;
 import frc.robot.subsystems.Index;
 import frc.robot.subsystems.Launcher;
 
 
-public class SemiAutoVersion4 extends CommandBase {
-   
-    public SemiAutoVersion4(){}
-    
-    public void excute(Index indexMotors, Launcher launcher){
-    // double tx = NetworkTableInstance.getDefault().getTable("limelight-back").getEntry("tx").getDouble(0);
-    double tx = 2; // Defining for testing purposes
-    // double ty = NetworkTableInstance.getDefault().getTable("limelight-back").getEntry("ty").getDouble(0);
-    
+public class SemiAutoVersion4 {
+  private static boolean targetLock = false;
+
+  static double tx = NetworkTableInstance.getDefault().getTable("limelight-back").getEntry("tx").getDouble(0);
+  
+  private static void targetStatus(Index indexMotors, Launcher launcher)  {  
       if (tx < 5) {
-          new IndexSpeed(indexMotors, 0.5);
-         } else {
-          new LaunchSpeed(launcher, 0.5, 0.5);}
+        targetLock = true;
+        return ; 
+      } // end of if conditional
+      
     }
 
-    public void end(boolean interrupted) {
-        ;
-      }
-
+  public void semiAutoLaunch(Index indexMotors, Launcher launcher){
+    
+    if (targetLock == true){
+      new IndexSpeed(indexMotors, 0.5);
+      } else {
+      new LaunchSpeed(launcher, 0.5, 0.5);}
+    }
 } // end of class
